# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Organisation_Rest_api",
        "description": "Organisation",
        "contact": {
            "email": "dev@organisation.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/v1/user/org_repository/add_repository": {
            "post": {
                "tags": [
                    "Repositories"
                ],
                "summary": "Create an organization repository",
                "description": "Fetches details from GitHub API and stores them in the database.",
                "operationId": "createOrgRepository",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "github_repo_url"
                                ],
                                "properties": {
                                    "github_repo_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://github.com/user/repo"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A sample repo"
                                    },
                                    "visibility": {
                                        "type": "string",
                                        "example": "public"
                                    },
                                    "language": {
                                        "type": "string",
                                        "example": "PHP"
                                    },
                                    "license": {
                                        "type": "string",
                                        "example": "MIT"
                                    },
                                    "forks_count": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "open_issues_count": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "watchers_count": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "default_branch": {
                                        "type": "string",
                                        "example": "main"
                                    },
                                    "topics": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "laravel",
                                            "vue",
                                            "devops"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Repository created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Repository created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to create repository"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Exception details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/user/org_repository/fetch": {
            "get": {
                "tags": [
                    "Repositories"
                ],
                "summary": "Fetch organization repositories",
                "description": "Retrieves a paginated list of repositories, optionally filtered by name or description.",
                "operationId": "fetchOrganisationRepo",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of repositories per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Search query to filter repositories by name or description",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "laravel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of repositories fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to fetch repositories."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Exception details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}